# everywhere: use utils::chunked_vector for list of mutations