… cql3/statements/create_aggregate_statement.hh
… cql3/statements/create_function_statement.cc
… cql3/statements/create_function_statement.hh
… cql3/statements/create_index_statement.cc
… cql3/statements/create_index_statement.hh
… cql3/statements/create_keyspace_statement.cc
… cql3/statements/create_table_statement.cc
… cql3/statements/create_table_statement.hh
… cql3/statements/create_type_statement.cc
… cql3/statements/create_type_statement.hh
… cql3/statements/create_view_statement.cc
… cql3/statements/create_view_statement.hh
… cql3/statements/drop_index_statement.cc
… cql3/statements/drop_index_statement.hh
… cql3/statements/drop_type_statement.cc
… cql3/statements/drop_type_statement.hh
… cql3/statements/drop_view_statement.cc
… cql3/statements/drop_view_statement.hh
… cql3/statements/modification_statement.cc
… cql3/statements/modification_statement.hh
… cql3/statements/schema_altering_statement.cc
… cql3/statements/schema_altering_statement.hh
… db/batchlog_manager.cc
… db/commitlog/commitlog.cc
… db/commitlog/commitlog.hh
… db/legacy_schema_migrator.cc
… db/schema_applier.cc
… db/schema_applier.hh
… db/schema_tables.cc
… db/schema_tables.hh