everywhere: use utils::chunked_vector for list of mutations
    
Currently, we use std::vector<*mutation> to keep
a list of mutations for processing.
This can lead to large allocation, e.g. when the vector
size is a function of the number of tables.

Use a chunked vector instead to prevent oversized allocations.

`perf-simple-query --smp 1` results obtained for fixed 400MHz frequency
and PGO disabled:

Before (read path):
```
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=read, query_single_key=no, counters=no}
Disabling auto compaction
Creating 10000 partitions...

89055.97 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39417 insns/op,   18003 cycles/op,        0 errors)
103372.72 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39380 insns/op,   17300 cycles/op,        0 errors)
98942.27 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39413 insns/op,   17336 cycles/op,        0 errors)
103752.93 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39407 insns/op,   17252 cycles/op,        0 errors)
102516.77 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39403 insns/op,   17288 cycles/op,        0 errors)
throughput:
        mean=   99528.13 standard-deviation=6155.71
        median= 102516.77 median-absolute-deviation=3844.59
        maximum=103752.93 minimum=89055.97
instructions_per_op:
        mean=   39403.99 standard-deviation=14.25
        median= 39406.75 median-absolute-deviation=9.30
        maximum=39416.63 minimum=39380.39
cpu_cycles_per_op:
        mean=   17435.81 standard-deviation=318.24
        median= 17300.40 median-absolute-deviation=147.59
        maximum=18002.53 minimum=17251.75
```

After (read path)
```
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=read, query_single_key=no, counters=no}
Disabling auto compaction
Creating 10000 partitions...
59755.04 tps ( 66.2 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39466 insns/op,   22834 cycles/op,        0 errors)
71854.16 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39417 insns/op,   17883 cycles/op,        0 errors)
82149.45 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.2 tasks/op,   39411 insns/op,   17409 cycles/op,        0 errors)
49640.04 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.3 tasks/op,   39474 insns/op,   19975 cycles/op,        0 errors)
54963.22 tps ( 66.1 allocs/op,   0.0 logallocs/op,  14.3 tasks/op,   39474 insns/op,   18235 cycles/op,        0 errors)
throughput:
        mean=   63672.38 standard-deviation=13195.12
        median= 59755.04 median-absolute-deviation=8709.16
        maximum=82149.45 minimum=49640.04
instructions_per_op:
        mean=   39448.38 standard-deviation=31.60
        median= 39466.17 median-absolute-deviation=25.75
        maximum=39474.12 minimum=39411.42
cpu_cycles_per_op:
        mean=   19267.01 standard-deviation=2217.03
        median= 18234.80 median-absolute-deviation=1384.25
        maximum=22834.26 minimum=17408.67
```

`perf-simple-query --smp 1 --write` results obtained for fixed 400MHz frequency
and PGO disabled:

Before (write path):
```
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=write, query_single_key=no, counters=no}
Disabling auto compaction
63736.96 tps ( 59.4 allocs/op,  16.4 logallocs/op,  14.3 tasks/op,   49667 insns/op,   19924 cycles/op,        0 errors)
64109.41 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   49992 insns/op,   20084 cycles/op,        0 errors)
56950.47 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   50005 insns/op,   20501 cycles/op,        0 errors)
44858.42 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   50014 insns/op,   21947 cycles/op,        0 errors)
28592.87 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   50027 insns/op,   27659 cycles/op,        0 errors)
throughput:
        mean=   51649.63 standard-deviation=15059.74
        median= 56950.47 median-absolute-deviation=12087.33
        maximum=64109.41 minimum=28592.87
instructions_per_op:
        mean=   49941.18 standard-deviation=153.76
        median= 50005.24 median-absolute-deviation=73.01
        maximum=50027.07 minimum=49667.05
cpu_cycles_per_op:
        mean=   22023.01 standard-deviation=3249.92
        median= 20500.74 median-absolute-deviation=1938.76
        maximum=27658.75 minimum=19924.32
```

After (write path)
```
enable-cache=1
Running test with config: {partitions=10000, concurrency=100, mode=write, query_single_key=no, counters=no}
Disabling auto compaction
53395.93 tps ( 59.4 allocs/op,  16.5 logallocs/op,  14.3 tasks/op,   50326 insns/op,   21252 cycles/op,        0 errors)
46527.83 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   50704 insns/op,   21555 cycles/op,        0 errors)
55846.30 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   50731 insns/op,   21060 cycles/op,        0 errors)
55669.30 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   50735 insns/op,   21521 cycles/op,        0 errors)
52130.17 tps ( 59.3 allocs/op,  16.0 logallocs/op,  14.3 tasks/op,   50757 insns/op,   21334 cycles/op,        0 errors)
throughput:
        mean=   52713.91 standard-deviation=3795.38
        median= 53395.93 median-absolute-deviation=2955.40
        maximum=55846.30 minimum=46527.83
instructions_per_op:
        mean=   50650.57 standard-deviation=182.46
        median= 50731.38 median-absolute-deviation=84.09
        maximum=50756.62 minimum=50325.87
cpu_cycles_per_op:
        mean=   21344.42 standard-deviation=202.86
        median= 21334.00 median-absolute-deviation=176.37
        maximum=21554.61 minimum=21060.24
```

Fixes #24815

Improvement for rare corner cases. No backport required