@avikivity 2025-07-10
In general chunked_vector just is a little worse than std::vector.

Please run perf-simple-query --smp 1, release mode, before and after (for read and write) and report results in the patch changelog.

Either disable pgo or refresh the pgo profiles before each build since it will have an effect on the fast path.