@mykaul 2025-07-10
CI passed in 2h? Wow.